+++ packages/vue-pose/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 Vue Pose adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.2] 2018-09-04
+
+### Removed
+
+- `preEnter` transition pending feedback.
+
+## [0.4.1] 2018-09-04
+
+### Fixed
+
+- Bug with initial pose on posed components within `PoseTransition`
+
+## [0.4.0] 2018-09-04
+
+### Added
+
+- `PoseTransition`.
+
 ## [0.3.0] 2018-09-03
 
 ### Fixes